treewide: Move all editing software into its own directory.
