+++ furios_gallery/edit_view/crop_overlay.py
