+++ furios_gallery/edit_view/draw_overlay.py
