+++ furios_gallery/edit_view/filters_overlay.py
