+++ furios_gallery/edit_view/image_transformations_overlay.py
